+++ shipPet/src/main/java/com/ship/controller/DepartmentController.java
@@ -53,7 +53,7 @@ public class DepartmentController {
 
     @DeleteMapping("/{id}")
     @Secured(RoleConst.ADMIN)
-    public Result delete(@PathVariable Long id) {
+    public Result delete(@PathVariable Integer id) {
         return Result.choice("删除单个", departService.deleteById(id));
     }
