+++ shipPet/src/main/java/com/ship/mapper/MsgMapper.java
@@ -22,30 +22,30 @@ public interface MsgMapper extends BaseMapper<Msg> {
     /**
      * 查询用户最近谈话的医生
      */
-    Long getByClient(@Param("clientId") Long clientId);
+    Integer getByClient(@Param("clientId") Integer clientId);
 
     /**
      * 查询特定用户与医生对话
      */
-    List<MsgVo> getADoctor(@Param("clientId") Long clientId, @Param("doctorId") Long doctorId);
+    List<MsgVo> getADoctor(@Param("clientId") Integer clientId, @Param("doctorId") Integer doctorId);
 
     /**
      * 查询与用户有过联系的医生
      */
-    List<MsgNameVo> getDoctorGroup(@Param("clientId") Long clientId, @Param("doctorId") Long doctorId);
+    List<MsgNameVo> getDoctorGroup(@Param("clientId") Integer clientId, @Param("doctorId") Integer doctorId);
 
     /**
      * 查询医生最近谈话的用户
      */
-    Long getByDoctor(@Param("doctorId") Long doctorId);
+    Integer getByDoctor(@Param("doctorId") Integer doctorId);
 
     /**
      * 查询特定医生与用户对话
      */
-    List<MsgVo> getAClient(@Param("doctorId") Long doctorId, @Param("clientId") Long clientId);
+    List<MsgVo> getAClient(@Param("doctorId") Integer doctorId, @Param("clientId") Integer clientId);
 
     /**
      * 医生谈过话的用户
      */
-    List<MsgNameVo> getClientGroup(@Param("doctorId") Long doctorId, @Param("clientId") Long clientId);
+    List<MsgNameVo> getClientGroup(@Param("doctorId") Integer doctorId, @Param("clientId") Integer clientId);
 }
