+++ shipPet/src/main/java/com/ship/vo/AdoptVo.java
@@ -3,5 +3,5 @@ package com.ship.vo;
 /**
  * 领养审核所需参数
  */
-public record AdoptVo(Long adoptId, Boolean isAdopt) {
+public record AdoptVo(Integer adoptId, Boolean isAdopt) {
 }
